Add share matrix logic

+++ src/misc.rs
@@ -2,7 +2,7 @@ use crate::field::{Fp, GADGET_VECTOR, L};
 use ff::{Field, PrimeFieldBits};
 use rand::Rng;
 
-pub fn rnd_fp_vec(size: u8, min: u64, max: u64) -> Vec<Fp> {
+pub fn rnd_fp_vec(size: usize, min: u64, max: u64) -> Vec<Fp> {
     let mut rng = rand::rng();
     (0..size).map(|_| Fp::from(rng.random_range(min..max))).collect()
 }
